did a little housekeeping
+in the create 'spells' I changed for loops so that ones that were [for
i in ___] in side of ANOTHER FOR LOOP USING I AS THE VARIABLE now use j
instead.
Common practice guys. Use it, I know you know how to because you have
examples in the same code.
+tried something with the createWorshipper spell, moving the spawned
setter into the if statement that checks if it's useable. Now minions
don't graphically spawn if they are unable to.